// frontend-angular/src/frontend-client/model/orderViewInitialData.ts
/**
 * Restaurant Import API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {GroupedOrderEntry} from './groupedOrderEntry';
import {DeliveryData} from './deliveryData';


export interface OrderViewInitialData {
    orderState: OrderViewInitialData.OrderStateEnum;
    orderDeliveryData: DeliveryData;
    restaurantName: string;
    restaurantTelephone: string;
    groupedEntries: Array<GroupedOrderEntry>;
    allEatingPeopleCount: number;
    basePriceSum: number;
    totalPrice: number;
}
export namespace OrderViewInitialData {
    export type OrderStateEnum = 'CREATED' | 'ORDERING' | 'ORDERED' | 'DELIVERED' | 'REJECTED';
    export const OrderStateEnum = {
        Created: 'CREATED' as OrderStateEnum,
        Ordering: 'ORDERING' as OrderStateEnum,
        Ordered: 'ORDERED' as OrderStateEnum,
        Delivered: 'DELIVERED' as OrderStateEnum,
        Rejected: 'REJECTED' as OrderStateEnum
    };
}